// generated/src/app/api-models/aucls-model-criteria.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuclsModelCriteria {
  auclsModelId: string;
  criteriaTypeCode: number;
  closedReasonCode: string;
  daysNo: number;
  criteriaAppliedFlag: string;
  securityCode: string;
  insertDateTime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}